src/test/java/com/leetcode/problems/rangeSumQuery2DImmutable/NumMatrixTest.java
package com.leetcode.problems.rangeSumQuery2DImmutable;

import org.junit.Test;

import static org.junit.Assert.*;

/**
 * @author Pavel Belevich
 */
public class NumMatrixTest {

    @Test
    public void sumRegionEmpty() throws Exception {
        new NumMatrix(new int[][]{{}});
    }

    @Test
    public void sumRegion() throws Exception {
        final NumMatrix numMatrix = new NumMatrix(new int[][]{
                {3, 0, 1, 4, 2},
                {5, 6, 3, 2, 1},
                {1, 2, 0, 1, 5},
                {4, 1, 0, 1, 7},
                {1, 0, 3, 0, 5}
        });
        assertEquals(8, numMatrix.sumRegion(2, 1, 4, 3));
        assertEquals(11, numMatrix.sumRegion(1, 1, 2, 2));
        assertEquals(12, numMatrix.sumRegion(1, 2, 2, 4));
    }

}